django_app/requirements.txt
django
django-environ
django-cors-headers
djangorestframework
psycopg2-binary
gdal
shapely 
networkx